requirements.txt
aiogram>=3.0.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
SQLModel>=0.0.8
APScheduler>=3.10.0
httpx>=0.24.0
pydantic>=2.0.0
twscrape>=0.17.0
loguru>=0.7.0
selenium>=4.15.0
webdriver-manager>=4.0.0
beautifulsoup4>=4.12.0
